# src/main/java/com/khoslak/WebMicroservice/model/CreateOrderRequest.java
package com.khoslak.WebMicroservice.model;


import com.fasterxml.jackson.annotation.JsonIdentityInfo;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CreateOrderRequest {
    private String item;
    private int quantity;
}