# src/main/java/com/khoslak/WebMicroservice/model/Order.java
package com.khoslak.WebMicroservice.model;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Order {
    private Long id;
    private String item;
    private int quantity;
    private String status;
}